accessible/tests/mochitest/text/test_passwords.html

<!DOCTYPE html>
<html>
<head>
  <title>nsIAccessibleText getText related function tests for text and password inputs</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../text.js"></script>

  <script type="application/javascript">
    function doTest() {
      // ////////////////////////////////////////////////////////////////////////
      // regular text and password inputs
      // ////////////////////////////////////////////////////////////////////////

      // //////////////////////////////////////////////////////////////////////
      // characterCount and getText for regular text field

      var IDs = [ "username" ];
      testCharacterCount(IDs, 4);
      testText(IDs, 0, 4, "test");

      // //////////////////////////////////////////////////////////////////////
      // characterCount and getText for password field

      IDs = [ "password" ];
      testCharacterCount(IDs, 4);
      testPasswordText(IDs, 0, 4, "test");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="mochitest for getText for password fields"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=415943">Mozilla Bug 415943</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <form action="post.php" method="post">
    <label for="username">User name:</label>
    <input id="username" value="test"><br />
    <label for="password">Password:</label>
    <input type="password" id="password" value="test"/>
  </form>
</body>
</html>
